+++ packages/showcases/react/stories/components/forms/VtmnSelect/VtmnSelect.stories.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
-import { VtmnSelect } from '@vtmn/react';
 import { Meta, Story } from '@storybook/react';
+import { VtmnSelect } from '@vtmn/react';
 import {
   argTypes,
   parameters,
 } from '@vtmn/showcase-core/csf/components/forms/select.csf';
+import React from 'react';
 
 export default {
   title: 'Components / Forms / VtmnSelect',
@@ -14,9 +14,15 @@ export default {
 } as Meta;
 
 const options = [
-  <option>option</option>,
-  <option>option 2</option>,
-  <option>option 3</option>,
+  <option value="" disabled selected>
+    Select
+  </option>,
+  <option value="gymnastics">Gymnastics</option>,
+  <option value="archery">Archery</option>,
+  <option value="climbing">Climbing</option>,
+  <option value="surf">Surf</option>,
+  <option value="badminton">Badminton</option>,
+  <option value="football">Football</option>,
 ];
 
 const Template: Story = (args) => <VtmnSelect options={options} {...args} />;