+++ packages/sources/vue/src/components/indicators/VtmnProgressbar/types.ts
@@ -1,0 +1,5 @@
+export type VtmnProgressbarVariant = 'linear' | 'circular';
+
+export type VtmnProgressbarSize = 'small' | 'medium' | 'large';
+
+export type VtmnProgressbarStatus = 'determinate' | 'indeterminate';