+++ packages/sources/vue/src/components/structure/VtmnListItem/types.ts
@@ -1,0 +1,1 @@
+export type VtmnListItemSize = 'small' | 'medium';