+++ packages/sources/css/CHANGELOG.md
@@ -3,6 +3,140 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.76.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.75.3...@vtmn/css@0.76.0) (2022-05-11)
+
+
+### Features
+
+* **@vtmn/css:** replace physical properties by logical properties ([#1140](https://github.com/Decathlon/vitamin-web/issues/1140)) ([71d13a8](https://github.com/Decathlon/vitamin-web/commit/71d13a8163fec6e3fc3c29647fbeadf46071b6ee))
+
+
+
+
+
+## [0.75.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.75.2...@vtmn/css@0.75.3) (2022-05-09)
+
+
+### Bug Fixes
+
+* **@vtmn/css-popover:** bottom position ([e307b3f](https://github.com/Decathlon/vitamin-web/commit/e307b3f2157353ad523eb64913870c9cbc629dfa))
+
+
+
+
+
+## [0.75.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.75.1...@vtmn/css@0.75.2) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
+## [0.75.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.75.0...@vtmn/css@0.75.1) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
+# [0.75.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.74.2...@vtmn/css@0.75.0) (2022-05-09)
+
+
+### Features
+
+* **@vtmn/css & @vtmn/vue & @vtmn/react & @vtmn/svelte:** chip, add role button + a11y attribute, improve css ([#1134](https://github.com/Decathlon/vitamin-web/issues/1134)) ([cef64c9](https://github.com/Decathlon/vitamin-web/commit/cef64c9a368824a9dd55bd0ad3a25e958d15aba3))
+
+
+
+
+
+## [0.74.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.74.1...@vtmn/css@0.74.2) (2022-05-06)
+
+
+### Bug Fixes
+
+* **@vtmn/css:** move flex auto to body ([#1150](https://github.com/Decathlon/vitamin-web/issues/1150)) ([f8ebfbf](https://github.com/Decathlon/vitamin-web/commit/f8ebfbf4f32b797c3dafa6caf457bd44c1ea56c0))
+
+
+
+
+
+## [0.74.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.74.0...@vtmn/css@0.74.1) (2022-05-05)
+
+
+### Bug Fixes
+
+* **@vtmn/svelte:** add flex auto + remove overflow indicator ([#1143](https://github.com/Decathlon/vitamin-web/issues/1143)) ([3082378](https://github.com/Decathlon/vitamin-web/commit/3082378ea90cae7c48e29a5f69c73afe6b568bcb))
+
+
+
+
+
+# [0.74.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.73.1...@vtmn/css@0.74.0) (2022-05-04)
+
+
+### Features
+
+* **@vtmn/css-quantity:** no default border ([#1141](https://github.com/Decathlon/vitamin-web/issues/1141)) ([bc8664c](https://github.com/Decathlon/vitamin-web/commit/bc8664c029a67d8d314faf259d7c83538e15f057))
+
+
+
+
+
+## [0.73.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.73.0...@vtmn/css@0.73.1) (2022-05-02)
+
+
+### Bug Fixes
+
+* **@vtmn/css:** edit `button` icon sizes, increase icon sizes for `icon-alone` variant, a11y  ([#1138](https://github.com/Decathlon/vitamin-web/issues/1138)) ([ecf7b25](https://github.com/Decathlon/vitamin-web/commit/ecf7b25799fd24f1b114837258b866032b4c5d6d))
+
+
+
+
+
+# [0.73.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.72.0...@vtmn/css@0.73.0) (2022-04-22)
+
+
+### Features
+
+* **@vtmn/css:** add screen readers support with `.sr-only` ([#1118](https://github.com/Decathlon/vitamin-web/issues/1118)) ([4a8ae38](https://github.com/Decathlon/vitamin-web/commit/4a8ae3825b50245813eacb88a88b06abc86e11a5))
+
+
+
+
+
+# [0.72.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.71.0...@vtmn/css@0.72.0) (2022-04-22)
+
+
+### Features
+
+* **@vtmn/css:** breadcrumb, add a11y aria-current ([#1130](https://github.com/Decathlon/vitamin-web/issues/1130)) ([01c033c](https://github.com/Decathlon/vitamin-web/commit/01c033c7eced1cf5fcd681cc1907856baa2766be))
+
+
+
+
+
+# [0.71.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.70.1...@vtmn/css@0.71.0) (2022-04-21)
+
+
+### Features
+
+* **@vtmn/css:** text-input & its showcases ([#1128](https://github.com/Decathlon/vitamin-web/issues/1128)) ([8c50684](https://github.com/Decathlon/vitamin-web/commit/8c5068470e97f8c542f336ea10aaf98c7070b1d5))
+
+
+
+
+
+## [0.70.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.70.0...@vtmn/css@0.70.1) (2022-04-21)
+
+**Note:** Version bump only for package @vtmn/css
+
+
+
+
+
 # [0.70.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css@0.69.0...@vtmn/css@0.70.0) (2022-04-14)