+++ packages/sources/css/src/components/forms/text-input/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.16.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-text-input@0.15.1...@vtmn/css-text-input@0.16.0) (2022-05-11)
+
+
+### Features
+
+* **@vtmn/css:** replace physical properties by logical properties ([#1140](https://github.com/Decathlon/vitamin-web/issues/1140)) ([71d13a8](https://github.com/Decathlon/vitamin-web/commit/71d13a8163fec6e3fc3c29647fbeadf46071b6ee))
+
+
+
+
+
+## [0.15.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-text-input@0.15.0...@vtmn/css-text-input@0.15.1) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-text-input
+
+
+
+
+
+# [0.15.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-text-input@0.14.1...@vtmn/css-text-input@0.15.0) (2022-04-21)
+
+
+### Features
+
+* **@vtmn/css:** text-input & its showcases ([#1128](https://github.com/Decathlon/vitamin-web/issues/1128)) ([8c50684](https://github.com/Decathlon/vitamin-web/commit/8c5068470e97f8c542f336ea10aaf98c7070b1d5))
+
+
+
+
+
+## [0.14.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-text-input@0.14.0...@vtmn/css-text-input@0.14.1) (2022-04-21)
+
+**Note:** Version bump only for package @vtmn/css-text-input
+
+
+
+
+
 # [0.14.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-text-input@0.13.14...@vtmn/css-text-input@0.14.0) (2022-04-08)