+++ packages/sources/css/src/components/navigation/breadcrumb/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.5.1...@vtmn/css-breadcrumb@0.6.0) (2022-05-11)
+
+
+### Features
+
+* **@vtmn/css:** replace physical properties by logical properties ([#1140](https://github.com/Decathlon/vitamin-web/issues/1140)) ([71d13a8](https://github.com/Decathlon/vitamin-web/commit/71d13a8163fec6e3fc3c29647fbeadf46071b6ee))
+
+
+
+
+
+## [0.5.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.5.0...@vtmn/css-breadcrumb@0.5.1) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-breadcrumb
+
+
+
+
+
+# [0.5.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.4.1...@vtmn/css-breadcrumb@0.5.0) (2022-04-22)
+
+
+### Features
+
+* **@vtmn/css:** breadcrumb, add a11y aria-current ([#1130](https://github.com/Decathlon/vitamin-web/issues/1130)) ([01c033c](https://github.com/Decathlon/vitamin-web/commit/01c033c7eced1cf5fcd681cc1907856baa2766be))
+
+
+
+
+
+## [0.4.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.4.0...@vtmn/css-breadcrumb@0.4.1) (2022-04-21)
+
+**Note:** Version bump only for package @vtmn/css-breadcrumb
+
+
+
+
+
 # [0.4.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-breadcrumb@0.3.16...@vtmn/css-breadcrumb@0.4.0) (2022-04-06)