+++ packages/sources/css/src/design-tokens/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.7](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-design-tokens@0.18.6...@vtmn/css-design-tokens@0.18.7) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-design-tokens
+
+
+
+
+
+## [0.18.6](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-design-tokens@0.18.5...@vtmn/css-design-tokens@0.18.6) (2022-04-21)
+
+**Note:** Version bump only for package @vtmn/css-design-tokens
+
+
+
+
+
 ## [0.18.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-design-tokens@0.18.4...@vtmn/css-design-tokens@0.18.5) (2022-03-11)
 
 **Note:** Version bump only for package @vtmn/css-design-tokens