+++ packages/sources/css/src/utilities/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.3.7...@vtmn/css-utilities@0.3.8) (2022-05-09)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
+## [0.3.7](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.3.6...@vtmn/css-utilities@0.3.7) (2022-04-21)
+
+**Note:** Version bump only for package @vtmn/css-utilities
+
+
+
+
+
 ## [0.3.6](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-utilities@0.3.5...@vtmn/css-utilities@0.3.6) (2022-03-11)
 
 **Note:** Version bump only for package @vtmn/css-utilities